修改安全告警问题

+++ frameworks/native/toneplayer/src/toneplayer_impl.cpp
@@ -292,7 +292,7 @@ int32_t TonePlayerImpl::GetSamples(uint16_t *freqs, int8_t *buffer, uint32_t req
         data = reinterpret_cast<uint8_t*>(buffer);
         double factor = freqVal * 2 * pi / samplingRate_; // 2 is a parameter in the sine wave formula
         for (uint32_t idx = 0; idx < reqSamples; idx++) {
-            int16_t sample = AMPLITUDE * sin(factor * index);
+            uint16_t sample = static_cast<uint16_t>(AMPLITUDE * sin(factor * index));
             uint32_t result;
             if (i == 0) {
                 result = (sample & 0xFF);
